pyproject.toml
[project]
name = "coding-agent-rlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

[tool.setuptools.packages.find]
include = ["functions*"]
